tests/wpt/web-platform-tests/offscreen-canvas/the-canvas-state/2d.state.saverestore.shadowOffsetX.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.state.saverestore.shadowOffsetX</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.state.saverestore.shadowOffsetX</h1>
<p class="desc">save()/restore() works for shadowOffsetX</p>


<script>
var t = async_test("save()/restore() works for shadowOffsetX");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

// Test that restore() undoes any modifications
var old = ctx.shadowOffsetX;
ctx.save();
ctx.shadowOffsetX = 5;
ctx.restore();
_assertSame(ctx.shadowOffsetX, old, "ctx.shadowOffsetX", "old");

// Also test that save() doesn't modify the values
ctx.shadowOffsetX = 5;
old = ctx.shadowOffsetX;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against 5
ctx.save();
_assertSame(ctx.shadowOffsetX, old, "ctx.shadowOffsetX", "old");
ctx.restore();

t.done();

});
</script>
